feat(notifiers):  :truck: move notifiers to own package and import it easier
