+++ flathunter/notifiers/__init__.py
@@ -1,0 +1,5 @@
+"""Package for notifiers."""
+from .sender_apprise import SenderApprise
+from .sender_mattermost import SenderMattermost
+from .sender_slack import SenderSlack
+from .sender_telegram import SenderTelegram
